pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "multi-fabric-data-agent-orchestration"
version = "0.1.0"
description = "Orchestrate multiple Microsoft Fabric data agents using AI-powered workflows"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = [
    "microsoft-fabric",
    "ai-agents",
    "langgraph",
    "agent-framework",
    "data-orchestration"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "azure-identity>=1.15.0",
    "azure-storage-file-datalake>=12.14.0",
    "requests>=2.31.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
]

[project.optional-dependencies]
agent-framework = [
    "agent-framework-azure-ai>=0.1.0",
]
langgraph = [
    "langchain>=0.2.0",
    "langchain-openai>=0.1.0",
    "langgraph>=0.2.0",
]
all = [
    "agent-framework-azure-ai>=0.1.0",
    "langchain>=0.2.0",
    "langchain-openai>=0.1.0",
    "langgraph>=0.2.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
]

[project.scripts]
fabric-agent-framework = "src.agent_framework.main:main"
fabric-langgraph = "src.langgraph.main:main"

[project.urls]
Homepage = "https://github.com/fuenfhausen/multi-fabric-data-agent-orchestration"
Repository = "https://github.com/fuenfhausen/multi-fabric-data-agent-orchestration"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.black]
line-length = 100
target-version = ["py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
